Add delete feature
For the moment it works only on single table query data visualization. No multitable query with functions are allowed. The 'delete row' button is visible also on this query visualization, maybe the solution should be to display it only for query where is possible to execute the delete of a row.